requirements.txt
streamlit>=1.28.0
langchain>=0.0.340
langchain-openai>=0.0.2
openai>=1.3.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pathlib2>=2.3.0
pydantic>=2.0.0
Pillow>=10.0.0
xlrd>=2.0.1
reportlab>=4.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
jsonschema>=4.20
networkx>=3.3
packaging>=24.0
pytest>=8.0.0
matplotlib>=3.8.0